sFTP scripted import — extract shared base RouteBuilder to dedupe SFTP/dynamic archiving+dispatch
ScriptedImportConversionDynamicRouteBuilder (REST) and
ScriptedImportConversionSftpDynamicRouteBuilder (SFTP polling) duplicated
their local-archive and item-dispatch logic almost verbatim
(archiveLocallyOnSuccess/OnError/archiveLocally, handleItemInList,
getErrorMessage, resolveCamelEndpointUri, ResponseAggregationStrategy).

Extract the shared behaviour into a new
AbstractScriptedImportConversionArchivingRouteBuilder that both extend.
The one real difference — how the archive file name is derived (the real
remote file name for SFTP vs. a synthesized name for REST, which has no
remote file) — is a Template Method: archiveFileName(Exchange).

Each subclass keeps only its own configure() (the routes genuinely
differ: direct: source vs. sftp: source, and the SFTP route's
onException(...).handled(true) to still consume/delete the remote file on
a failed transform). Also drops the redundant @Slf4j logger on the SFTP
builder in favour of the SLF4J logger RouteBuilder already provides
(matching the sibling enable/disable route builders in this package).
